change line in Ch 7 so people don't unwittingly mess up a var
This is a fix for a problem I encountered this with a participant while coaching during the 3/28/2020 virtual class. Right now, the instructions for "Filtering a DataFrame" in Chapter 7 include a line of code showing a pandas filter expression.

```
Now, in the next cell we will form the filter expression expected by pandas.

`committee_list.prop_name == my_prop`
```

If you aren't paying close attention and you _run_ that line of code, you end up messing with your `committee_list` variable and then the next steps don't work. In the live version of the class, as people follow along, it works fine because Ben doesn't execute this line separately—and clearly it's printed here in the docs just to highlight what the filter expression is supposed to look like.

But for people catching up or following the scripted class, it's easy to overlook that the docs don't say to run this line. If they do, they can get stuck when their `my_committees` filtered variable later on has 0 entries.